:recycle: Extract PillSheet
